+++ src/library/pkgcache/NEWS.md
@@ -1,5 +1,9 @@
 # pkgcache (development version)
 
+* All HTTP requests now honor the `pkgcache_http_version` option and
+  `PKGCACHE_HTTP_VERSION` environment variable. Closes
+  https://github.com/r-lib/pkgcache/issues/140.
+
 * New `PKG_USE_BIOCONDUCTOR` environment variable and new
   `pkg.use_bioconductor` option to opt out from automatic Bioconductor
   support.